One function per file in interchange
