+++ packages/lodestar-validator/src/slashingProtection/interchange/types.ts
@@ -1,4 +1,9 @@
 import {BLSPubkey, Root, SlashingProtectionAttestation, SlashingProtectionBlock} from "@chainsafe/lodestar-types";
+import {IInterchangeCompleteV4} from "./formats/completeV4";
+
+export type Interchange = IInterchangeCompleteV4;
+
+export type InterchangeFormatVersion = {format: "complete"; version: "4"};
 
 export interface IInterchangeLodestar {
   genesisValidatorsRoot: Root;
